+++ system-parachains/people/people-kusama/src/tests.rs
@@ -14,110 +14,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-	coretime::{BrokerPalletId, CoretimeBurnAccount},
-	xcm_config::LocationToAccountId,
-	*,
-};
-use coretime::CoretimeAllocator;
-use frame_support::{
-	assert_ok,
-	traits::{
-		fungible::{Inspect, Mutate},
-		Get, OnInitialize,
-	},
-};
-use kusama_runtime_constants::system_parachain::coretime::TIMESLICE_PERIOD;
-use pallet_broker::{ConfigRecordOf, RCBlockNumberOf, SaleInfo};
-use parachains_runtimes_test_utils::ExtBuilder;
+use crate::xcm_config::LocationToAccountId;
+use polkadot_primitives::AccountId;
 use sp_core::crypto::Ss58Codec;
-use sp_runtime::traits::AccountIdConversion;
+use xcm::prelude::*;
 use xcm_runtime_apis::conversions::LocationToAccountHelper;
 
 const ALICE: [u8; 32] = [1u8; 32];
 
-fn advance_to(b: BlockNumber) {
-	while System::block_number() < b {
-		let block_number = System::block_number() + 1;
-		System::set_block_number(block_number);
-		Broker::on_initialize(block_number);
-	}
-}
-
-#[test]
-fn bulk_revenue_is_burnt() {
-	ExtBuilder::<Runtime>::default()
-		.with_collators(vec![AccountId::from(ALICE)])
-		.with_session_keys(vec![(
-			AccountId::from(ALICE),
-			AccountId::from(ALICE),
-			SessionKeys { aura: AuraId::from(sp_core::sr25519::Public::from_raw(ALICE)) },
-		)])
-		.build()
-		.execute_with(|| {
-			// Configure broker and start sales
-			let config = ConfigRecordOf::<Runtime> {
-				advance_notice: 1,
-				interlude_length: 1,
-				leadin_length: 2,
-				region_length: 1,
-				ideal_bulk_proportion: Perbill::from_percent(100),
-				limit_cores_offered: None,
-				renewal_bump: Perbill::from_percent(3),
-				contribution_timeout: 1,
-			};
-			let ed = ExistentialDeposit::get();
-			assert_ok!(Broker::configure(RuntimeOrigin::root(), config.clone()));
-			assert_ok!(Broker::start_sales(RuntimeOrigin::root(), ed, 1));
-
-			let sale_start = SaleInfo::<Runtime>::get().unwrap().sale_start;
-			advance_to(sale_start + config.interlude_length);
-
-			// Check and set initial balances.
-			let broker_account = BrokerPalletId::get().into_account_truncating();
-			let coretime_burn_account = CoretimeBurnAccount::get();
-			let treasury_account = xcm_config::RelayTreasuryPalletAccount::get();
-			assert_ok!(Balances::mint_into(&AccountId::from(ALICE), 200 * ed));
-			let alice_balance_before = Balances::balance(&AccountId::from(ALICE));
-			let treasury_balance_before = Balances::balance(&treasury_account);
-			let broker_balance_before = Balances::balance(&broker_account);
-			let burn_balance_before = Balances::balance(&coretime_burn_account);
-
-			// Purchase coretime.
-			assert_ok!(Broker::purchase(RuntimeOrigin::signed(AccountId::from(ALICE)), 100 * ed));
-
-			// Alice decreases.
-			assert!(Balances::balance(&AccountId::from(ALICE)) < alice_balance_before);
-			// Treasury balance does not increase.
-			assert_eq!(Balances::balance(&treasury_account), treasury_balance_before);
-			// Broker pallet account does not increase.
-			assert_eq!(Balances::balance(&broker_account), broker_balance_before);
-			// Coretime burn pot gets the funds.
-			assert!(Balances::balance(&coretime_burn_account) > burn_balance_before);
-
-			// They're burnt when a day has passed on chain.
-			// This needs to be asserted in an emulated test.
-		});
-}
-
-#[test]
-fn timeslice_period_is_sane() {
-	// Config TimeslicePeriod is set to this constant - assumption in burning logic.
-	let timeslice_period_config: RCBlockNumberOf<CoretimeAllocator> =
-		<Runtime as pallet_broker::Config>::TimeslicePeriod::get();
-	assert_eq!(timeslice_period_config, TIMESLICE_PERIOD);
-
-	// Timeslice period constant non-zero - assumption in burning logic.
-	#[cfg(feature = "fast-runtime")]
-	assert_eq!(TIMESLICE_PERIOD, 20);
-	#[cfg(not(feature = "fast-runtime"))]
-	assert_eq!(TIMESLICE_PERIOD, 80);
-}
-
 #[test]
 fn location_conversion_works() {
-	let alice_32 =
-		AccountId32 { network: None, id: polkadot_core_primitives::AccountId::from(ALICE).into() };
+	let alice_32 = AccountId32 { network: None, id: AccountId::from(ALICE).into() };
 	let bob_20 = AccountKey20 { network: None, key: [123u8; 20] };
 
 	// the purpose of hardcoded values is to catch an unintended location conversion logic change.
@@ -203,13 +110,13 @@ fn location_conversion_works() {
 	];
 
 	for tc in test_cases {
-		let expected = polkadot_core_primitives::AccountId::from_string(tc.expected_account_id_str)
-			.expect("Invalid AccountId string");
+		let expected =
+			AccountId::from_string(tc.expected_account_id_str).expect("Invalid AccountId string");
 
-		let got = LocationToAccountHelper::<polkadot_core_primitives::AccountId, LocationToAccountId>::convert_location(
+		let got = LocationToAccountHelper::<AccountId, LocationToAccountId>::convert_location(
 			tc.location.into(),
 		)
-			.unwrap();
+		.unwrap();
 
 		assert_eq!(got, expected, "{}", tc.description);
 	}